[CpmlArc] Using size_t instead of int where appropriate
The n_curves arg is better expressed by size_t.

+++ cpml/cpml-arc.h
@@ -41,7 +41,7 @@ void    cpml_arc_to_cairo               (const CpmlPrimitive    *arc,
                                          cairo_t                *cr);
 void    cpml_arc_to_curves              (const CpmlPrimitive    *arc,
                                          CpmlSegment            *segment,
-                                         int                     n_curves);
+                                         size_t                  n_curves);
 
 CAIRO_END_DECLS
